Integrate cleans up unless arg is passed [full ci]

+++ packages/azure/start_emulators.ps1
@@ -8,7 +8,7 @@ function Log-Time
 
 $StartTime = Get-Date
 
-if ($env:RUNNING_IN_CI -eq 1) 
+if ($env:RUNNING_IN_CI -eq 1)
 {
     Write-Host "Running in CI"
     # Env Vars required and set in CI (not required in the Windows VM but required when running intergration tests locally)
@@ -37,15 +37,22 @@ Write-Host "COSMOS_SECRET_KEY: $env:COSMOS_SECRET_KEY"
 $protocol = [Net.ServicePointManager]::SecurityProtocol
 Write-Host "Protocol: $protocol"
 
-$d1 = Get-Date
-Write-Host "Downloading Cosmos Emulator"
-curl https://aka.ms/cosmosdb-emulator -o .\cosmos.msi
-Log-Time -Start $d1
+if (Test-Path -Path .\cosmos.msi -PathType leaf)
+{
+    Write-Host "Cosmos Emulator already downloaded and installed"
+}
+else
+{
+    $d1 = Get-Date
+    Write-Host "Downloading Cosmos Emulator"
+    curl https://aka.ms/cosmosdb-emulator -o .\cosmos.msi
+    Log-Time -Start $d1
 
-$d1 = Get-Date
-Write-Host "Installing Cosmos Emulator"
-Start-Process -Wait .\cosmos.msi -ArgumentList "/quiet"
-Log-Time -Start $d1
+    $d1 = Get-Date
+    Write-Host "Installing Cosmos Emulator"
+    Start-Process -Wait .\cosmos.msi -ArgumentList "/quiet"
+    Log-Time -Start $d1
+}
 
 Write-Host "Loading CosmosDB Module"
 Import-Module "$env:ProgramFiles\Azure Cosmos DB Emulator\PSModules\Microsoft.Azure.CosmosDB.Emulator"
@@ -58,48 +65,62 @@ Log-Time -Start $d1
 $status = Get-CosmosDbEmulatorStatus
 Write-Host "Current Cosmos Emulator Status: $status"
 
-$d1 = Get-Date
-Write-Host "Downloading SqlLocalDB"
-curl 'https://download.microsoft.com/download/8/D/D/8DD7BDBA-CEF7-4D8E-8C16-D9F69527F909/ENU/x64/SqlLocalDB.MSI' -o .\SqlLocalDB.MSI
-Log-Time -Start $d1
-
-$d1 = Get-Date
-Write-Host "Installing SqlLocalDB"
-Start-Process -Wait .\SqlLocalDB.MSI -ArgumentList "/qn","IACCEPTSQLLOCALDBLICENSETERMS=YES"
-Log-Time -Start $d1
-
-$d1 = Get-Date
-Write-Host "Downloading Storage Emulator"
-curl 'https://go.microsoft.com/fwlink/?linkid=717179&clcid=0x409' -o .\az_storage_emulator.msi
-Log-Time -Start $d1
-
-$d1 = Get-Date
-Write-Host "Installing Storage Emulator"
-Start-Process -Wait .\az_storage_emulator.msi -ArgumentList "/quiet"
-Log-Time -Start $d1
-
-if ($env:RUNNING_IN_CI -ne 1) 
+if (Test-Path -Path .\SqlLocalDB.MSI -PathType leaf)
 {
-    $vm_ip = (Get-NetIPAddress -InterfaceAlias "Ethernet" -AddressFamily "IPv4").IPAddress
-    Write-Host "Configure IP from 127.0.0.1 to $vm_ip for Windows VM"
-    $storage_emulator_config_path = "C:\Program Files (x86)\Microsoft SDKs\Azure\Storage Emulator\AzureStorageEmulator.exe.config"
-    ((Get-Content -path $storage_emulator_config_path -Raw) -replace '127.0.0.1', $vm_ip ) | Set-Content -Path $storage_emulator_config_path
+    Write-Host "SqlLocalDB already downloaded and installed"
 }
+else
+{
+    $d1 = Get-Date
+    Write-Host "Downloading SqlLocalDB"
+    curl 'https://download.microsoft.com/download/8/D/D/8DD7BDBA-CEF7-4D8E-8C16-D9F69527F909/ENU/x64/SqlLocalDB.MSI' -o .\SqlLocalDB.MSI
+    Log-Time -Start $d1
 
-$d1 = Get-Date
-Write-Host "Create MSSQLLocalDB"
-Start-Process -Wait "C:\Program Files\Microsoft SQL Server\110\Tools\Binn\SqlLocalDB.exe" -ArgumentList "create","MSSQLLocalDB"
-Log-Time -Start $d1
+    $d1 = Get-Date
+    Write-Host "Installing SqlLocalDB"
+    Start-Process -Wait .\SqlLocalDB.MSI -ArgumentList "/qn","IACCEPTSQLLOCALDBLICENSETERMS=YES"
+    Log-Time -Start $d1
+
+    $d1 = Get-Date
+    Write-Host "Create MSSQLLocalDB"
+    Start-Process -Wait "C:\Program Files\Microsoft SQL Server\110\Tools\Binn\SqlLocalDB.exe" -ArgumentList "create","MSSQLLocalDB"
+    Log-Time -Start $d1
+}
 
 $d1 = Get-Date
 Write-Host "Start MSSQLLocalDB"
 Start-Process "C:\Program Files\Microsoft SQL Server\110\Tools\Binn\SqlLocalDB.exe" -ArgumentList "start","MSSQLLocalDB"
 Log-Time -Start $d1
 
-$d1 = Get-Date
-Write-Host  "Initializing Storage Emulator"
-Start-Process -Wait "C:\Program Files (x86)\Microsoft SDKs\Azure\Storage Emulator\AzureStorageEmulator.exe"  -ArgumentList "init","/server","(localdb)\.\MSSQLLocalDb","-inprocess"
-Log-Time -Start $d1
+if (Test-Path -Path .\az_storage_emulator.msi -PathType leaf)
+{
+    Write-Host "Storage Emulator already downloaded and installed"
+}
+else
+{
+    $d1 = Get-Date
+    Write-Host "Downloading Storage Emulator"
+    curl 'https://go.microsoft.com/fwlink/?linkid=717179&clcid=0x409' -o .\az_storage_emulator.msi
+    Log-Time -Start $d1
+
+    $d1 = Get-Date
+    Write-Host "Installing Storage Emulator"
+    Start-Process -Wait .\az_storage_emulator.msi -ArgumentList "/quiet"
+    Log-Time -Start $d1
+
+    if ($env:RUNNING_IN_CI -ne 1) 
+    {
+        $vm_ip = (Get-NetIPAddress -InterfaceAlias "Ethernet" -AddressFamily "IPv4").IPAddress
+        Write-Host "Configure IP from 127.0.0.1 to $vm_ip for Windows VM"
+        $storage_emulator_config_path = "C:\Program Files (x86)\Microsoft SDKs\Azure\Storage Emulator\AzureStorageEmulator.exe.config"
+        ((Get-Content -path $storage_emulator_config_path -Raw) -replace '127.0.0.1', $vm_ip ) | Set-Content -Path $storage_emulator_config_path
+    }
+
+    $d1 = Get-Date
+    Write-Host  "Initializing Storage Emulator"
+    Start-Process -Wait "C:\Program Files (x86)\Microsoft SDKs\Azure\Storage Emulator\AzureStorageEmulator.exe"  -ArgumentList "init","/server","(localdb)\.\MSSQLLocalDb","-inprocess"
+    Log-Time -Start $d1
+}
 
 $d1 = Get-Date
 Write-Host "Starting Storage Emulator"
